% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{call_api}
\alias{call_api}
\alias{fopi_get_year}
\alias{fopi_get_country}
\alias{fopi_get_rank}
\title{Call API}
\usage{
fopi_get_year(year)

fopi_get_country(country)

fopi_get_rank(rank)
}
\arguments{
\item{year}{Year of data to fetch.}

\item{country}{Country of data to fetch.}

\item{rank}{Rank of data to fetch.}
}
\description{
Get data via the API.
}
