% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fopi}
\alias{fopi}
\title{Freedom of Press Index}
\format{A data.frame (\code{tibble}) object with 3111 rows and 5 columns
\describe{
  \item{\code{country}}{ Country}
  \item{\code{country_code}}{ ISO2c country code}
  \item{\code{year}}{ Year of index}
  \item{\code{rank}}{ Freedom of Press rank}
  \item{\code{score}}{ Freedom of Press score}
}}
\source{
\url{https://en.wikipedia.org/wiki/Press_Freedom_Index}
}
\usage{
fopi
}
\description{
Tidy Freedom of press index data from Wikipedia.
}
\keyword{datasets}
