% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{run_api}
\alias{run_api}
\title{Run API}
\usage{
run_api(host = "0.0.0.0", port = 3000)
}
\arguments{
\item{host, port}{Parameters passed to \link[plumber]{plumb}.}
}
\description{
Run the Freedom of Press Index API.
}
